Fixes issue #18
Have now added a check to see if there were already video files.
Old video files will be removed if new ones are copied. A message
box will inform the user.

Checked that the interaction works as intended.

+++ video_selector.py
@@ -1,6 +1,7 @@
+import os
 import shutil
 import tkinter as tk
-from tkinter import filedialog
+from tkinter import filedialog,messagebox
 
 from dtrack_params import dtrack_params
 from project import project_file
@@ -117,20 +118,42 @@ class VideoSelector(tk.Toplevel):
         track_video = self.__slt_tracking_video.get_entry_information()
 
         if self.__blv_copy.get():
-            # Preserve filename
-            calib_video_filename = calib_video.split("/")[-1]
-            track_video_filename = track_video.split("/")[-1]
+            # Get current video filenames, check if the files exist, and if so
+            # warn the user that copying videos locally will remove the old ones.
+            old_calib_video = project_file["calibration_video"]
+            old_track_video = project_file["tracking_video"]
+
+            calib_exists = os.path.exists(old_calib_video)
+            track_exists = os.path.exists(old_track_video)
+
+            delete_old_videos = False
+            if calib_exists or track_exists:
+                msg = "This project directory already contains one or both video files. " +\
+                      " Copying new video files will overwrite the old ones. Do you wish to continue?"
+                
+                delete_old_videos = messagebox.askyesno(parent=self,
+                                                        title="Overwrite old videos?",
+                                                        message=msg)
+
+                if not delete_old_videos:
+                    # User cancelled or clicked 
+                    return
+
+            calib_video_filename = os.path.basename(calib_video)
+            track_video_filename = os.path.basename(track_video)
 
 
             # Given that we have the project file record, why rename these at 
             # all? We don't have to assume video filenames anymore
             local_calib_name =\
-                 dtrack_params['project_directory'] + "/" + calib_video_filename
-                      
-            local_track_name =\
-                  dtrack_params['project_directory'] + "/" + track_video_filename
+                os.path.join(dtrack_params["project_directory"], calib_video_filename)
 
-            # Copy video files locally
+            local_track_name =\
+                os.path.join(dtrack_params["project_directory"], track_video_filename)
+
+            # Copy video files locally, if the user has tried to copy a file to
+            # itself, just ignore it. There are legitimate cases where this can
+            # happen (e.g. updating one video but not the other)
             try:
                 shutil.copy(calib_video, local_calib_name)    
             except shutil.SameFileError:
@@ -153,6 +176,19 @@ class VideoSelector(tk.Toplevel):
             # Overwrite so that project file points to local copies
             calib_video = local_calib_name
             track_video = local_track_name
+
+
+            # If the user opted to delete old videos, check they aren't identical
+            # to the newly selected videos and delete. 
+            if delete_old_videos:
+                if not (calib_video == old_calib_video):
+                    print("Removing old calibration video at: {}".format(old_calib_video))
+                    os.remove(old_calib_video)
+                if not (track_video == old_track_video):
+                    print("Removing old tracking video at: {}".format(old_track_video))
+                    os.remove(old_track_video)
+
+            
         
         project_file["calibration_video"] = calib_video
         project_file["tracking_video"] = track_video
